% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mic.R
\name{annotate_mic}
\alias{annotate_mic}
\title{Create annotate csv}
\usage{
annotate_mic(path, output)
}
\arguments{
\item{path}{directory which contains mic100's csv files}

\item{output}{csv annotate file}
}
\description{
Create annotate csv
}
