% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic100-package.R
\docType{package}
\name{mic100-package}
\alias{mic100}
\alias{mic100-package}
\title{mic100: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
