% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mic.R
\name{summarise_mic}
\alias{summarise_mic}
\title{Create MIC-100 summary csv}
\usage{
summarise_mic(path, output)
}
\arguments{
\item{path}{directory which contains mic100's csv files}

\item{output}{name of summary file}
}
\description{
Create MIC-100 summary csv
}
