% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mic100-package.R
\docType{package}
\name{mic100-package}
\alias{mic100}
\alias{mic100-package}
\title{mic100: Tools for MIC-100}
\description{
Read and manage photosynthetic gas-exchange data logged by MIC-100
}
\author{
\strong{Maintainer}: Keach Murakami \email{keach.murakami@gmail.com} (0000-0001-8150-9535)

}
\keyword{internal}
