% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mic.R
\encoding{UTF-8}
\name{read_mic}
\alias{read_mic}
\title{Read Photosynthetic Rate Logged by MIC100}
\usage{

}
\arguments{
\item{file}{path to the csv file}
}
\description{
Read Photosynthetic Rate Logged by MIC100
}
