% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mic.R
\name{read_mic}
\alias{read_mic}
\title{Read Photosynthetic gas-exchange parameters Logged by MIC-100}
\usage{
read_mic(file, simple = TRUE)
}
\arguments{
\item{file}{path to the csv file}
}
\description{
Read Photosynthetic gas-exchange parameters Logged by MIC-100
}
